Added Lexer, Token, Errors; Appended Constants
The Lexer class was added - it tokenizes the text that is inputed.

Token class was added - Used within the Lexer class for the tokenization
to hold information about the code inputed

Added base Error class - Has as_string method to make it easy to read.

Added IllegalCharError class - Used within the Lexer class, should a
character that shouldn't be given as an input, Lexer will raise this
error. Inherits from Error base class.

Appended the Constants file to add the float token type, as well as the
alphanumeric constants.