Added Lexer, Token, Errors; Updated CMakeLists in Library
Lexer class was added - Tokenizes the text that is inputted.

Token class was added - Used within the Lexer class for the tokenization
to hold the information of the text inputted

Base Error class was added - has toString() function to make it easy to
read the error

IllegalCharacterError class was added - Used within the Lexer, if a
character is in the inputted text that isn't supposed to be, the Lexer
will raise this error

Updated CMakeLists to include all the header files and the cpp files for
the newly added classes.